Move new fixtures to conftest.py
