+++ test/csp_helpers.py
@@ -13,7 +13,6 @@ from unittest import mock
 from urllib.parse import parse_qs, urlparse
 
 import jwt
-import pytest
 from botocore.awsrequest import AWSRequest
 from botocore.credentials import Credentials
 
@@ -307,34 +306,3 @@ class FakeAwsEnvironment:
     def __exit__(self, *args, **kwargs):
         for patcher in self.patchers:
             patcher.__exit__(*args, **kwargs)
-
-
-@pytest.fixture
-def no_metadata_service():
-    """Emulates an environment without any metadata service."""
-    with NoMetadataService() as server:
-        yield server
-
-
-@pytest.fixture
-def fake_aws_environment():
-    """Emulates the AWS environment, returning dummy credentials."""
-    with FakeAwsEnvironment() as env:
-        yield env
-
-
-@pytest.fixture(
-    params=[FakeAzureFunctionMetadataService(), FakeAzureVmMetadataService()],
-    ids=["azure_function", "azure_vm"],
-)
-def fake_azure_metadata_service(request):
-    """Parameterized fixture that emulates both the Azure VM and Azure Functions metadata services."""
-    with request.param as server:
-        yield server
-
-
-@pytest.fixture
-def fake_gce_metadata_service():
-    """Emulates the GCE metadata service, returning a dummy token."""
-    with FakeGceMetadataService() as server:
-        yield server
